Revisions Based on Esri Comments
This upload adds two features per the requests of Geoff Taylor at ESRI. The first is that a person will be placed in front of the door for scale if a parameter is switched on, and a height limit exceedance is calculated based on a height limit parameter. An exceeded is both emitted as a report and has a scenario representation that can be viewed. In the future, I intend to make sure it can change the building so an existing 3D building can be used instead. The same code can likely apply for this feature in the building shell rule/something else.